Revert "updating develop to the latest geoman version "
