+++ .github/workflows/publish.yml
@@ -1,0 +1,43 @@
+name: publish-package
+
+on: release
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/setup-node@v1
+        with:
+          node-version: 12
+      - run: npm ci
+      - run: npm test
+
+  publish-npm:
+    needs: build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/setup-node@v1
+        with:
+          node-version: 12
+          registry-url: https://registry.npmjs.org/
+      - run: npm ci
+      - run: npm publish
+        env:
+          NODE_AUTH_TOKEN: ${{secrets.NPM}}
+
+  # publish-gpr:
+    # needs: build
+    # runs-on: ubuntu-latest
+    # steps:
+    #   - uses: actions/checkout@v1
+    #   - uses: actions/setup-node@v1
+    #     with:
+    #       node-version: 12
+    #       registry-url: https://npm.pkg.github.com/
+    #       scope: '@codeofsumit'
+    #   - run: npm ci
+    #   - run: npm publish
+    #     env:
+    #       NODE_AUTH_TOKEN: ${{secrets.GPR}}
