+++ CONTRIBUTING.md
@@ -1,12 +1,11 @@
 ## Contributing
 
-Thank you for considering contributing to Leaflet-Geoman.
+Thank you for considering contributing to leaflet.pm.
 
 Follow these steps to get up and running:
 
 1.  clone the repository
 2.  run `npm install`
 3.  run `npm start`
-4.  run `npm run cypress` to launch the test suite (also nice for TDD)
-5. Launch developing.spec.js, it will open Chrome with a map and leaflet.pm where you can test the changes you make, manually and with the test suite
-6. In case you need to write tests for your new feature, create a separate *.spec.js file for your tests or add them to an existing one when appropriate
+4.  open `index.html` in your browser
+5.  run `npm run cypress` to launch the test suite (also nice for TDD).
