--- cypress/fixtures/PolygonPart1.json
@@ -1,96 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {
-        "shape": "Polygon",
-        "maps": [
-          "default-map"
-        ],
-        "name": "Unnamed Layer",
-        "category": "default",
-        "id": "fa4cf31e-38f2-4b00-b1ff-43ccb3b2439e"
-      },
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -74.012189,
-              40.719713
-            ],
-            [
-              -74.012833,
-              40.720884
-            ],
-            [
-              -74.012704,
-              40.7216
-            ],
-            [
-              -74.012575,
-              40.72238
-            ],
-            [
-              -74.012446,
-              40.723909
-            ],
-            [
-              -74.012918,
-              40.72495
-            ],
-            [
-              -74.014678,
-              40.724527
-            ],
-            [
-              -74.015365,
-              40.723616
-            ],
-            [
-              -74.015365,
-              40.721209
-            ],
-            [
-              -74.014463,
-              40.719323
-            ],
-            [
-              -74.013047,
-              40.717957
-            ],
-            [
-              -74.009914,
-              40.7162
-            ],
-            [
-              -74.004507,
-              40.715745
-            ],
-            [
-              -74.001846,
-              40.718607
-            ],
-            [
-              -74.000773,
-              40.720656
-            ],
-            [
-              -74.002361,
-              40.724137
-            ],
-            [
-              -74.011116,
-              40.72495
-            ],
-            [
-              -74.012189,
-              40.719713
-            ]
-          ]
-        ]
-      }
-    }
-  ]
-}
