--- cypress/fixtures/PolygonPart2.json
@@ -1,200 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {
-        "shape": "Polygon",
-        "maps": [
-          "default-map"
-        ],
-        "name": "Unnamed Layer",
-        "category": "default",
-        "id": "7af7d4e3-0111-4119-bff6-465c5e004481"
-      },
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -74.011116,
-              40.72495
-            ],
-            [
-              -74.002361,
-              40.724137
-            ],
-            [
-              -74.000773,
-              40.720656
-            ],
-            [
-              -74.001846,
-              40.718607
-            ],
-            [
-              -74.004507,
-              40.715745
-            ],
-            [
-              -74.009914,
-              40.7162
-            ],
-            [
-              -74.013991,
-              40.71594
-            ],
-            [
-              -74.014163,
-              40.715322
-            ],
-            [
-              -74.013863,
-              40.714574
-            ],
-            [
-              -74.012189,
-              40.714574
-            ],
-            [
-              -74.011588,
-              40.715192
-            ],
-            [
-              -74.010043,
-              40.714867
-            ],
-            [
-              -74.009957,
-              40.714509
-            ],
-            [
-              -74.009957,
-              40.714281
-            ],
-            [
-              -74.009314,
-              40.714021
-            ],
-            [
-              -74.008241,
-              40.714118
-            ],
-            [
-              -74.007125,
-              40.714346
-            ],
-            [
-              -74.005666,
-              40.714216
-            ],
-            [
-              -74.00485,
-              40.713923
-            ],
-            [
-              -74.001846,
-              40.713956
-            ],
-            [
-              -74.001675,
-              40.714346
-            ],
-            [
-              -74.001675,
-              40.714671
-            ],
-            [
-              -74.00146,
-              40.715127
-            ],
-            [
-              -74.000473,
-              40.715712
-            ],
-            [
-              -73.998413,
-              40.716135
-            ],
-            [
-              -73.996267,
-              40.716395
-            ],
-            [
-              -73.995752,
-              40.716818
-            ],
-            [
-              -73.995709,
-              40.717046
-            ],
-            [
-              -73.996568,
-              40.718607
-            ],
-            [
-              -73.996611,
-              40.719388
-            ],
-            [
-              -73.996139,
-              40.720299
-            ],
-            [
-              -73.995709,
-              40.721209
-            ],
-            [
-              -73.995709,
-              40.72264
-            ],
-            [
-              -73.998284,
-              40.723974
-            ],
-            [
-              -73.999443,
-              40.724657
-            ],
-            [
-              -74.000258,
-              40.725437
-            ],
-            [
-              -74.000473,
-              40.725958
-            ],
-            [
-              -74.00043,
-              40.726934
-            ],
-            [
-              -74.002233,
-              40.726836
-            ],
-            [
-              -74.003005,
-              40.726055
-            ],
-            [
-              -74.004722,
-              40.72534
-            ],
-            [
-              -74.008198,
-              40.72612
-            ],
-            [
-              -74.008284,
-              40.727226
-            ],
-            [
-              -74.011116,
-              40.72495
-            ]
-          ]
-        ]
-      }
-    }
-  ]
-}
