--- cypress/integration/developing.spec.js
@@ -1,7 +1,0 @@
-describe('Opens Testing Environment', () => {
-  const mapSelector = '#map';
-
-  it('opens a map in a browser', () => {
-
-  });
-});
