+++ cypress/integration/marker.spec.js
@@ -1,29 +1,6 @@
 describe('Draw Marker', () => {
   const mapSelector = '#map';
 
-  it('enables drag in programatic global edit mode', () => {
-
-    cy.toolbarButton('marker').click();
-
-    cy.wait(1000)
-
-    cy.get(mapSelector)
-      .click(150, 250)
-
-    cy.wait(1000)
-
-    cy.window().then(({ map, L }) => {
-      map.pm.enableGlobalEditMode()
-
-      map.eachLayer((layer) => {
-        if (layer instanceof L.Marker) {
-          assert.isTrue(layer.dragging._enabled)
-        }
-      })
-    });
-
-  });
-
   it('removes markers without error', () => {
     cy.window().then(({ map, L }) => {
       const markerLayer = L.geoJson().addTo(map);
