+++ demo/index-canvas.html
@@ -1,46 +1,43 @@
 <!DOCTYPE>
 <html>
+    <head>
+        <title>Leaflet Polygon Management Demo</title>
+        <link rel="stylesheet" href="https://unpkg.com/leaflet@1.2.0/dist/leaflet.css" />
+        <link rel="stylesheet" href="../dist/leaflet.pm.css">
+        <link rel="stylesheet" href="demo.css" />
+        <script src="https://unpkg.com/leaflet@1.2.0/dist/leaflet.js"></script>
+        <!-- <script src="../../leaflet-measure-path/leaflet-measure-path.js"></script> -->
+        <!-- <script src="https://unpkg.com/leaflet.markercluster@1.0.0"></script> -->
 
-<head>
-    <title>Leaflet Polygon Management Demo</title>
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@latest/dist/leaflet.css" />
-    <link rel="stylesheet" href="../dist/leaflet-geoman.css">
-    <link rel="stylesheet" href="demo.css" />
-    <script src="https://unpkg.com/leaflet@latest/dist/leaflet.js"></script>
-    <!-- <script src="../../leaflet-measure-path/leaflet-measure-path.js"></script> -->
-    <!-- <script src="https://unpkg.com/leaflet.markercluster@1.0.0"></script> -->
+    </head>
+    <body>
 
-</head>
+        <div class="wrapper">
 
-<body>
+            <article>
+                <h1>Leaflet Polygon Management</h1>
+                <h3>An Open Source Leaflet Plugin for editing polygons, specifically made for Leaflet 1.0</h3>
+            </article>
+            <article>
+                <h2>Creating A Polygon</h2>
 
-    <div class="wrapper">
+                <div class="map" id="example2"></div>
 
-        <article>
-            <h1>Leaflet Polygon Management</h1>
-            <h3>An Open Source Leaflet Plugin for editing polygons, specifically made for Leaflet 1.0</h3>
-        </article>
-        <article>
-            <h2>Creating A Polygon</h2>
+            </article>
+            <article>
+                <h2>Editing A Polygon</h2>
 
-            <div class="map" id="example2"></div>
+                <div class="map" id="example3"></div>
 
-        </article>
-        <article>
-            <h2>Editing A Polygon</h2>
+            </article>
+            <article>
+                <h2>Editing A Layer Group</h2>
 
-            <div class="map" id="example3"></div>
+                <div class="map" id="example4"></div>
 
-        </article>
-        <article>
-            <h2>Editing A Layer Group</h2>
-
-            <div class="map" id="example4"></div>
-
-        </article>
-    </div>
-    <script src="../dist/leaflet-geoman.min.js"></script>
-    <script src="demo-canvas.js"></script>
-</body>
-
+            </article>
+        </div>
+        <script src="../dist/leaflet.pm.min.js"></script>
+        <script src="demo-canvas.js"></script>
+    </body>
 </html>
