+++ demo/index.html
@@ -5,9 +5,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Leaflet Polygon Management Demo</title>
     <link rel="stylesheet" href="https://unpkg.com/leaflet@1.2.0/dist/leaflet.css" />
-    <link rel="stylesheet" href="../dist/leaflet-geoman.css">
+    <link rel="stylesheet" href="../dist/leaflet.pm.css">
     <link rel="stylesheet" href="demo.css" />
-    <script src="https://unpkg.com/leaflet@latest/dist/leaflet.js"></script>
+    <script src="https://unpkg.com/leaflet@1.3.1/dist/leaflet.js"></script>
     <!-- <script src="../../leaflet-measure-path/leaflet-measure-path.js"></script> -->
     <!-- <script src="https://unpkg.com/leaflet.markercluster@1.0.0"></script> -->
 
@@ -40,7 +40,7 @@
 
         </article>
     </div>
-    <script src="../dist/leaflet-geoman.min.js"></script>
+    <script src="../dist/leaflet.pm.min.js"></script>
     <script src="demo.js"></script>
 </body>
