+++ index.html
@@ -1,50 +1,48 @@
-<!DOCTYPE>
+<!DOCTYPE >
 <html>
-
-<head>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Leaflet Geometry Management</title>
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@latest/dist/leaflet.css" />
-    <link rel="stylesheet" href="dist/leaflet-geoman.css" />
-    <script src="https://unpkg.com/leaflet@latest/dist/leaflet.js"></script>
-    <!--
+    <head>
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <title>Leaflet Geometry Management</title>
+        <link
+            rel="stylesheet"
+            href="https://unpkg.com/leaflet@latest/dist/leaflet.css"
+        />
+        <link rel="stylesheet" href="dist/leaflet.pm.css" />
+        <script src="https://unpkg.com/leaflet@latest/dist/leaflet.js"></script>
+        <!--
             <script src="../../leaflet-measure-path/leaflet-measure-path.js"></script>
         -->
-    <!--
+        <!--
             <script src="https://unpkg.com/leaflet.markercluster@1.0.0"></script>
         -->
 
-    <style>
-        html,
-        body {
-            font-family: Arial, Verdana;
-            background-color: #eef1ef;
-            font-size: 16px;
-            width: 100%;
-            height: 100%;
-            margin: 0;
-        }
+        <style>
+            html,
+            body {
+                font-family: Arial, Verdana;
+                background-color: #eef1ef;
+                font-size: 16px;
+                width: 100%;
+                height: 100%;
+                margin: 0;
+            }
 
-        .wrapper {
-            width: 100%;
-            height: 100%;
-        }
+            .wrapper {
+                width: 100%;
+                height: 100%;
+            }
 
-        #map {
-            height: 500px;
-        }
-    </style>
-</head>
+            #map {
+                height: 500px;
+            }
+        </style>
+    </head>
 
-<body>
-    <div class="wrapper">
-        <div id="map"></div>
-    </div>
-    <!--
-            <script src="https://unpkg.com/@geoman-io/leaflet-geoman-free@latest/dist/leaflet-geoman.min.js"></script>
-
+    <body>
+        <div class="wrapper"><div id="map"></div></div>
+        <!--
+            <script src="https://unpkg.com/leaflet.pm@latest/dist/leaflet.pm.min.js"></script>
         -->
-    <script src="dist/leaflet-geoman.min.js"></script>
-</body>
-
+        <script src="dist/leaflet.pm.min.js"></script>
+    </body>
 </html>
