+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@geoman-io/leaflet-geoman-free",
-  "version": "2.5.0",
+  "name": "leaflet.pm",
+  "version": "2.2.0",
   "description": "A Leaflet Plugin For Editing Geometry Layers in Leaflet 1.0",
   "keywords": [
     "leaflet",
@@ -17,7 +17,7 @@
   "files": [
     "dist"
   ],
-  "main": "dist/leaflet-geoman.min.js",
+  "main": "dist/leaflet.pm.min.js",
   "dependencies": {
     "@turf/difference": "5.1.5",
     "@turf/intersect": "^6.1.3",
@@ -25,24 +25,24 @@
     "lodash": "^4.17.15"
   },
   "devDependencies": {
-    "@babel/core": "^7.8.4",
-    "@babel/preset-env": "^7.8.4",
+    "@babel/core": "^7.6.0",
+    "@babel/preset-env": "^7.6.0",
     "babel-loader": "^8.0.6",
-    "css-loader": "^3.4.2",
-    "cypress": "^3.8.3",
+    "css-loader": "^2.1.1",
+    "cypress": "^3.4.1",
     "eslint": "^4.19.1",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-config-prettier": "^3.6.0",
-    "eslint-plugin-cypress": "^2.9.0",
-    "eslint-plugin-import": "^2.20.1",
-    "file-loader": "^4.3.0",
-    "leaflet": "^1.6.0",
-    "mini-css-extract-plugin": "^0.8.2",
-    "prettier": "^1.19.1",
-    "uglifyjs-webpack-plugin": "^2.2.0",
-    "url-loader": "^2.3.0",
-    "webpack": "^4.41.5",
-    "webpack-cli": "^3.3.10"
+    "eslint-plugin-cypress": "^2.6.1",
+    "eslint-plugin-import": "^2.18.2",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^0.11.1",
+    "leaflet": "^1.5.1",
+    "prettier": "1.16.1",
+    "style-loader": "^0.19.0",
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.12.0"
   },
   "peerDependencies": {
     "leaflet": "^1.2.0"
@@ -50,8 +50,8 @@
   "scripts": {
     "start": "npm run dev",
     "dev": "./node_modules/.bin/webpack --config=webpack.dev.js",
-    "test": "./node_modules/.bin/cypress run",
-    "cypress": "./node_modules/.bin/cypress open",
+    "test": "$(npm bin)/cypress run",
+    "cypress": "$(npm bin)/cypress open",
     "build": "./node_modules/.bin/webpack --config=webpack.build.js",
     "prepare": "npm run build",
     "eslint-check": "eslint --print-config . | eslint-config-prettier-check",
@@ -61,7 +61,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git://github.com/geoman-io/leaflet-geoman.git"
+    "url": "git://github.com/codeofsumit/leaflet.pm.git"
   },
   "author": {
     "name": "Sumit Kumar",
@@ -70,9 +70,9 @@
   },
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/geoman-io/leaflet-geoman/issues"
+    "url": "https://github.com/codeofsumit/leaflet.pm/issues"
   },
-  "homepage": "https://geoman.io/leaflet-geoman",
+  "homepage": "https://leafletpm.now.sh",
   "prettier": {
     "trailingComma": "es5",
     "tabWidth": 2,
