+++ src/assets/translations/en.json
@@ -25,8 +25,6 @@
     "dragButton": "Drag Layers",
     "cutButton": "Cut Layers",
     "deleteButton": "Remove Layers",
-    "drawCircleMarkerButton": "Draw Circle Marker",
-    "snappingButton": "Snap dragged marker to other layers and vertices",
-    "pinningButton": "Pin shared vertices together"
+    "drawCircleMarkerButton": "Draw Circle Marker"
   }
 }
