+++ src/css/controls.css
@@ -1,5 +1,4 @@
 .leaflet-pm-toolbar {
-  
 }
 
 .leaflet-pm-toolbar .leaflet-buttons-control-button {
@@ -7,19 +6,6 @@
   box-sizing: border-box;
   position: relative;
   z-index: 3;
-}
-
-.leaflet-pm-toolbar a:first-child {
-  border-radius: 0;
-}
-
-.leaflet-pm-toolbar .button-container:first-child a {
-  border-top-left-radius: 4px;
-  border-top-right-radius: 4px;
-}
-.leaflet-pm-toolbar .button-container:last-child a {
-  border-bottom-left-radius: 4px;
-  border-bottom-right-radius: 4px;
 }
 
 .leaflet-pm-toolbar .control-fa-icon {
@@ -66,9 +52,6 @@
 .leaflet-pm-toolbar .leaflet-pm-icon-cut {
   background-image: url('../assets/icons/Scissors.svg');
 }
-.leaflet-pm-toolbar .leaflet-pm-icon-snapping {
-  background-image: url('../assets/icons/Magnet.svg');
-}
 
 .leaflet-buttons-control-button:hover {
   cursor: pointer;
@@ -90,7 +73,7 @@
   z-index: 2;
   position: absolute;
   top: 0;
-  left: 100%;
+  left: 31px;
   display: none;
   white-space: nowrap;
 }
