+++ src/js/Draw/L.PM.Draw.Cut.js
@@ -15,7 +15,7 @@ Draw.Cut = Draw.Polygon.extend({
     const layers = Object.keys(all)
       // convert object to array
       .map(l => all[l])
-      // only layers handled by leaflet-geoman
+      // only layers handled by leaflet.pm
       .filter(l => l.pm)
       // only polygons
       .filter(l => l instanceof L.Polygon)
@@ -36,24 +36,12 @@ Draw.Cut = Draw.Polygon.extend({
       const diff = difference(l.toGeoJSON(15), layer.toGeoJSON(15));
 
       // the resulting layer after the cut
-      const resultingLayer = L.geoJSON(diff, l.options).addTo(this._map);
+      const resultingLayer = L.geoJSON(diff, l.options);
 
       // give the new layer the original options
       resultingLayer.pm.enable(this.options);
       resultingLayer.pm.disable();
 
-      // add templayer prop so pm:remove isn't fired
-      l._pmTempLayer = true;
-      layer._pmTempLayer = true;
-
-      // remove old layer and cutting layer
-      l.remove();
-      layer.remove();    
-
-      if (resultingLayer.getLayers().length === 0) {
-        this._map.pm.removeLayer({ target: resultingLayer });
-      }   
-               
       let resultingLayers = [resultingLayer];
 
       if (
@@ -76,8 +64,6 @@ Draw.Cut = Draw.Polygon.extend({
         });
       }
 
-
-
       // fire pm:cut on the cutted layer
       l.fire('pm:cut', {
         shape: this._shape,
@@ -92,7 +78,17 @@ Draw.Cut = Draw.Polygon.extend({
         cuttedLayer: l,
       });
 
+      // add templayer prop so pm:remove isn't fired
+      l._pmTempLayer = true;
+      layer._pmTempLayer = true;
 
+      // remove old layer and cutting layer
+      l.remove();
+      layer.remove();
+
+      if (resultingLayer.getLayers().length === 0) {
+        this._map.pm.removeLayer({ target: resultingLayer });
+      }
     });
   },
   _finishShape() {
