+++ src/js/Draw/L.PM.Draw.js
@@ -19,15 +19,12 @@ const Draw = L.Class.extend({
       draggable: true,
     },
   },
-  setOptions(options) {
-    L.Util.setOptions(this, options);
-  },
   initialize(map) {
     // save the map
     this._map = map;
 
-    // define all possible shapes that can be drawn
-    this.shapes = ['Marker', 'CircleMarker', 'Line', 'Polygon', 'Rectangle', 'Circle', 'Cut'];
+        // define all possible shapes that can be drawn
+        this.shapes = ['Marker', 'CircleMarker', 'Line', 'Polygon', 'Rectangle', 'Circle', 'Cut'];
 
     // initiate drawing class for our shapes
     this.shapes.forEach(shape => {
