+++ src/js/Edit/L.PM.Edit.CircleMarker.js
@@ -5,24 +5,6 @@ Edit.CircleMarker = Edit.extend({
     this._layer = layer;
     this._enabled = false;
   },
-  applyOptions() {
-    if (this.options.draggable) {
-      this.enableLayerDrag();
-    } else {
-      this.disableLayerDrag();
-    }
-
-    if (this.options.snappable) {
-      this._initSnappableMarkers();
-    } else {
-      this._disableSnapping();
-    }
-
-    // enable removal for the marker
-    if (!this.options.preventMarkerRemoval) {
-      this._layer.on('contextmenu', this._removeMarker, this);
-    }
-  },
   toggleEdit(options) {
     if (!this.enabled()) {
       this.enable(options);
@@ -49,33 +31,46 @@ Edit.CircleMarker = Edit.extend({
       this.disable();
     }
 
-    this.applyOptions();
-
     // change state
     this._enabled = true;
+
+    // enable removal for the marker
+    if (!this.options.preventMarkerRemoval) {
+      this._layer.on('contextmenu', this._removeMarker, this);
+    }
+
+    if (this.options.draggable) {
+      this.enableLayerDrag();
+    }
+
+    if (this.options.snappable) {
+      this._initSnappableMarkers();
+    }
 
     this._layer.on('pm:dragend', this._onMarkerDragEnd, this);
   },
   disable(layer = this._layer) {
+    // if it's not enabled, it doesn't need to be disabled
+    if (!this.enabled()) {
+      return false;
+    }
+
     // prevent disabling if layer is being dragged
     if (layer.pm._dragging) {
       return false;
     }
+    layer.pm._enabled = false;
 
-    // disable dragging, as this could have been active even without being enabled
-    this.disableLayerDrag();
-
-    // only fire events if it was enabled before
-    if (!this.enabled()) {
-      this._layer.fire('pm:disable');
-
-      if (this._layerEdited) {
-        this._layer.fire('pm:update', {});
-      }
-      this._layerEdited = false;
+    // remove draggable class
+    if (layer._path) {
+      const el = layer._path;
+      L.DomUtil.removeClass(el, 'leaflet-pm-draggable');
     }
 
-    layer.pm._enabled = false;
+    if (this._layerEdited) {
+      this._layer.fire('pm:update', {});
+    }
+    this._layerEdited = false;
 
     return true;
   },
@@ -114,11 +109,4 @@ Edit.CircleMarker = Edit.extend({
     marker.off('pm:dragstart', this._unsnap, this);
     marker.on('pm:dragstart', this._unsnap, this);
   },
-  _disableSnapping() {
-    const marker = this._layer;
-
-    marker.off('pm:drag', this._handleSnapping, this);
-    marker.off('pm:dragend', this._cleanupSnapping, this);
-    marker.off('pm:dragstart', this._unsnap, this);
-  }
 });
