+++ src/js/Edit/L.PM.Edit.js
@@ -1,5 +1,5 @@
 import SnapMixin from '../Mixins/Snapping';
-import DragMixin from '../Mixins/Dragging';
+import DragMixin from '../Mixins/Drag';
 
 const Edit = L.Class.extend({
   includes: [DragMixin, SnapMixin],
@@ -8,12 +8,7 @@ const Edit = L.Class.extend({
     snapDistance: 20,
     allowSelfIntersection: true,
     draggable: true,
-    limitMarkersToCount: -1
   },
-  setOptions(options) {
-    L.Util.setOptions(this, options);
-  },
-  applyOptions() { },
   isPolygon() {
     // if it's a polygon, it means the coordinates array is multi dimensional
     return this._layer instanceof L.Polygon;
