+++ src/js/L.PM.Utils.js
@@ -7,27 +7,6 @@ const Utils = {
 
     return map.unproject(p1._add(p2)._divideBy(2));
   },
-  findLayers(map) {
-    let layers = [];
-    map.eachLayer(layer => {
-      if (
-        layer instanceof L.Polyline ||
-        layer instanceof L.Marker ||
-        layer instanceof L.Circle ||
-        layer instanceof L.CircleMarker
-      ) {
-        layers.push(layer);
-      }
-    });
-
-    // filter out layers that don't have the leaflet-geoman instance
-    layers = layers.filter(layer => !!layer.pm);
-
-    // filter out everything that's leaflet-geoman specific temporary stuff
-    layers = layers.filter(layer => !layer._pmTempLayer);
-
-    return layers;
-  },
 };
 
 export default Utils;
