+++ src/js/Mixins/Drag.js
@@ -1,8 +1,5 @@
 const DragMixin = {
   enableLayerDrag() {
-    // before enabling layer drag, disable layer editing
-    this.disable();
-
     if (this._layer instanceof L.Marker) {
       this._layer.dragging.enable();
       return;
