+++ src/js/Mixins/Snapping.js
@@ -9,9 +9,6 @@ const SnapMixin = {
     this._layer.off('pm:dragstart', this._unsnap, this);
     this._layer.on('pm:dragstart', this._unsnap, this);
   },
-  _disableSnapping() {
-    this._layer.off('pm:dragstart', this._unsnap, this);
-  },
   _assignEvents(markerArr) {
     // loop through marker array and assign events to the markers
     markerArr.forEach(marker => {
@@ -63,10 +60,6 @@ const SnapMixin = {
     this._snapList.splice(index, 1);
   },
   _handleSnapping(e) {
-    function throttledList() {
-      return L.Util.throttle(this._createSnapList, 100, this);
-    }
-
     // if snapping is disabled via holding ALT during drag, stop right here
     if (e.originalEvent.altKey) {
       return false;
@@ -76,11 +69,7 @@ const SnapMixin = {
     // this isn't inside a movestart/dragstart callback because middlemarkers are initialized
     // after dragstart/movestart so it wouldn't fire for them
     if (this._snapList === undefined) {
-      this._createSnapList();
-
-      // re-create the snaplist again when a layer is added during draw
-      this._map.off('layeradd', throttledList, this);
-      this._map.on('layeradd', throttledList, this);
+      this._createSnapList(e);
     }
 
     // if there are no layers to snap to, stop here
@@ -265,7 +254,7 @@ const SnapMixin = {
       layer => layer._latlng || (layer._latlngs && layer._latlngs.length > 0)
     );
 
-    // finally remove everything that's leaflet-geoman specific temporary stuff
+    // finally remove everything that's leaflet.pm specific temporary stuff
     layers = layers.filter(layer => !layer._pmTempLayer);
 
     // save snaplist from layers and the other snap layers added from other classes/scripts
