+++ src/js/Toolbar/L.Controls.js
@@ -6,19 +6,15 @@ const PMButton = L.Control.extend({
   },
   // TODO: clean up variable names like _button should be _options and that domNodeVariable stuff
   initialize(options) {
-    // replaced setOptions with this because classNames returned undefined 🤔
-    this._button = Object.assign({}, this.options, options);
+    this._button = L.Util.setOptions(this, options);
   },
   onAdd(map) {
     this._map = map;
 
-    if (this._button.tool === 'edit') {
-      this._container = this._map.pm.Toolbar.editContainer;
-    } else if (this._button.tool === 'options') {
-      this._container = this._map.pm.Toolbar.optionsContainer;
-    } else {
-      this._container = this._map.pm.Toolbar.drawContainer;
-    }
+    this._container =
+      this._button.tool === 'edit'
+        ? this._map.pm.Toolbar.editContainer
+        : this._map.pm.Toolbar.drawContainer;
     this.buttonsDomNode = this._makeButton(this._button);
     this._container.appendChild(this.buttonsDomNode);
 
@@ -91,12 +87,6 @@ const PMButton = L.Control.extend({
           this._triggerClick();
         },
       },
-      finishMode: {
-        text: getTranslation('actions.finish'),
-        onClick() {
-          this._triggerClick();
-        },
-      },
       removeLastVertex: {
         text: getTranslation('actions.removeLastVertex'),
         onClick() {
@@ -126,7 +116,7 @@ const PMButton = L.Control.extend({
     });
 
     if (button.toggleStatus) {
-      L.DomUtil.addClass(buttonContainer, 'active');
+      L.DomUtil.addClass(newButton, 'active');
     }
 
     const image = L.DomUtil.create('div', 'control-icon', newButton);
