+++ src/js/Toolbar/L.PM.Toolbar.js
@@ -16,7 +16,6 @@ const Toolbar = L.Class.extend({
     dragMode: true,
     cutPolygon: true,
     removalMode: true,
-    snappingOption: true,
     position: 'topleft',
   },
   initialize(map) {
@@ -44,10 +43,6 @@ const Toolbar = L.Class.extend({
     this.editContainer = L.DomUtil.create(
       'div',
       'leaflet-pm-toolbar leaflet-pm-edit leaflet-bar leaflet-control'
-    );
-    this.optionsContainer = L.DomUtil.create(
-      'div',
-      'leaflet-pm-toolbar leaflet-pm-options leaflet-bar leaflet-control'
     );
 
     this._defineButtons();
@@ -132,14 +127,8 @@ const Toolbar = L.Class.extend({
     // we can't have two active modes because of possible event conflicts
     // so, we trigger a click on all currently active (toggled) buttons
 
-    // the options toolbar should not be disabled during the different modes
-    // TODO: probably need to abstract this a bit so different options are automatically
-    // disabled for different modes, like pinning for circles
-    const exceptOptionButtons = ['snappingOption']
-
     for (const name in this.buttons) {
       if (
-        !exceptOptionButtons.includes(name) &&
         this.buttons[name] !== exceptThisButton &&
         this.buttons[name].toggled()
       ) {
@@ -147,7 +136,7 @@ const Toolbar = L.Class.extend({
       }
     }
   },
-  toggleButton(name, status, disableOthers = true) {
+  toggleButton(name, status) {
     // does not fire the events/functionality of the button
     // this just changes the state and is used if a functionality (like Draw)
     // is enabled manually via script
@@ -162,9 +151,7 @@ const Toolbar = L.Class.extend({
 
     // as some mode got enabled, we still have to trigger the click on the other buttons
     // to disable their mode
-    if (disableOthers) {
-      this.triggerClickOnToggledButtons(this.buttons[name]);
-    }
+    this.triggerClickOnToggledButtons(this.buttons[name]);
 
     // now toggle the state of the button
     return this.buttons[name].toggle(status);
@@ -175,7 +162,7 @@ const Toolbar = L.Class.extend({
       className: 'control-icon leaflet-pm-icon-marker',
       title: getTranslation('buttonTitles.drawMarkerButton'),
       jsClass: 'Marker',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         // toggle drawing mode
         this.map.pm.Draw.Marker.toggle();
@@ -191,7 +178,7 @@ const Toolbar = L.Class.extend({
       title: getTranslation('buttonTitles.drawPolyButton'),
       className: 'control-icon leaflet-pm-icon-polygon',
       jsClass: 'Polygon',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         // toggle drawing mode
         this.map.pm.Draw.Polygon.toggle();
@@ -207,7 +194,7 @@ const Toolbar = L.Class.extend({
       className: 'control-icon leaflet-pm-icon-polyline',
       title: getTranslation('buttonTitles.drawLineButton'),
       jsClass: 'Line',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         // toggle drawing mode
         this.map.pm.Draw.Line.toggle();
@@ -223,7 +210,7 @@ const Toolbar = L.Class.extend({
       title: getTranslation('buttonTitles.drawCircleButton'),
       className: 'control-icon leaflet-pm-icon-circle',
       jsClass: 'Circle',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         // toggle drawing mode
         this.map.pm.Draw.Circle.toggle();
@@ -239,7 +226,7 @@ const Toolbar = L.Class.extend({
       title: getTranslation('buttonTitles.drawCircleMarkerButton'),
       className: 'control-icon leaflet-pm-icon-circle-marker',
       jsClass: 'CircleMarker',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         // toggle drawing mode
         this.map.pm.Draw.CircleMarker.toggle();
@@ -255,7 +242,7 @@ const Toolbar = L.Class.extend({
       title: getTranslation('buttonTitles.drawRectButton'),
       className: 'control-icon leaflet-pm-icon-rectangle',
       jsClass: 'Rectangle',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         // toggle drawing mode
         this.map.pm.Draw.Rectangle.toggle();
@@ -270,7 +257,7 @@ const Toolbar = L.Class.extend({
     const editButton = {
       title: getTranslation('buttonTitles.editButton'),
       className: 'control-icon leaflet-pm-icon-edit',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         this.map.pm.toggleGlobalEditMode();
       },
@@ -279,13 +266,13 @@ const Toolbar = L.Class.extend({
       disableOtherButtons: true,
       position: this.options.position,
       tool: 'edit',
-      actions: ['finishMode'],
+      actions: ['cancel'],
     };
 
     const dragButton = {
       title: getTranslation('buttonTitles.dragButton'),
       className: 'control-icon leaflet-pm-icon-drag',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         this.map.pm.toggleGlobalDragMode();
       },
@@ -294,14 +281,14 @@ const Toolbar = L.Class.extend({
       disableOtherButtons: true,
       position: this.options.position,
       tool: 'edit',
-      actions: ['finishMode'],
+      actions: ['cancel'],
     };
 
     const cutButton = {
       title: getTranslation('buttonTitles.cutButton'),
       className: 'control-icon leaflet-pm-icon-cut',
       jsClass: 'Cut',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         // enable polygon drawing mode without snap
         this.map.pm.Draw.Cut.toggle({
@@ -321,7 +308,7 @@ const Toolbar = L.Class.extend({
     const deleteButton = {
       title: getTranslation('buttonTitles.deleteButton'),
       className: 'control-icon leaflet-pm-icon-delete',
-      onClick: () => { },
+      onClick: () => {},
       afterClick: () => {
         this.map.pm.toggleGlobalRemovalMode();
       },
@@ -330,7 +317,7 @@ const Toolbar = L.Class.extend({
       disableOtherButtons: true,
       position: this.options.position,
       tool: 'edit',
-      actions: ['finishMode'],
+      actions: ['cancel'],
     };
 
     this._addButton('drawMarker', new L.Control.PMButton(drawMarkerButton));
