+++ src/js/polyfills.js
@@ -24,7 +24,7 @@ Array.prototype.findIndex =
   };
 
 // Array.find Polyfill for IE<12.
-// Requested here: https://github.com/geoman-io/leaflet-geoman/issues/173
+// Requested here: https://github.com/codeofsumit/leaflet.pm/issues/173
 Array.prototype.find =
   Array.prototype.find ||
   function(callback) {
@@ -86,56 +86,3 @@ if (typeof Object.assign != 'function') {
     });
   });
 })([Element.prototype, CharacterData.prototype, DocumentType.prototype]);
-
-// Polyfill for Array.includes()
-// https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Global_Objects/Array/includes#Polyfill
-if (!Array.prototype.includes) {
-  Object.defineProperty(Array.prototype, 'includes', {
-    value: function(searchElement, fromIndex) {
-
-      if (this == null) {
-        throw new TypeError('"this" is null or not defined');
-      }
-
-      // 1. Let O be ? ToObject(this value).
-      var o = Object(this);
-
-      // 2. Let len be ? ToLength(? Get(O, "length")).
-      var len = o.length >>> 0;
-
-      // 3. If len is 0, return false.
-      if (len === 0) {
-        return false;
-      }
-
-      // 4. Let n be ? ToInteger(fromIndex).
-      //    (If fromIndex is undefined, this step produces the value 0.)
-      var n = fromIndex | 0;
-
-      // 5. If n ≥ 0, then
-      //  a. Let k be n.
-      // 6. Else n < 0,
-      //  a. Let k be len + n.
-      //  b. If k < 0, let k be 0.
-      var k = Math.max(n >= 0 ? n : len - Math.abs(n), 0);
-
-      function sameValueZero(x, y) {
-        return x === y || (typeof x === 'number' && typeof y === 'number' && isNaN(x) && isNaN(y));
-      }
-
-      // 7. Repeat, while k < len
-      while (k < len) {
-        // a. Let elementK be the result of ? Get(O, ! ToString(k)).
-        // b. If SameValueZero(searchElement, elementK) is true, return true.
-        if (sameValueZero(o[k], searchElement)) {
-          return true;
-        }
-        // c. Increase k by 1. 
-        k++;
-      }
-
-      // 8. Return false
-      return false;
-    }
-  });
-}
