+++ webpack.dev.js
@@ -1,24 +1,18 @@
 /* eslint import/no-extraneous-dependencies: 0 */
 
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const path = require('path');
 
 module.exports = {
     watch: true,
     devtool: 'cheap-eval-source-map',
-    mode: 'development',
     entry: ['./src/js/L.PM.js'],
     output: {
-        filename: 'leaflet-geoman.min.js',
+        filename: 'leaflet.pm.min.js',
         path: path.resolve(__dirname, 'dist'),
     },
     module: {
         rules: [
-            {
-                test: /\.mjs$/,
-                include: /node_modules/,
-                type: "javascript/auto",
-            },
             {
                 test: /\.js$/,
                 exclude: /(node_modules|bower_components)/,
@@ -31,9 +25,10 @@ module.exports = {
             },
             {
                 test: /\.css$/,
-                use: [{
-                    loader: MiniCssExtractPlugin.loader,
-                }, 'css-loader',],
+                use: ExtractTextPlugin.extract({
+                    fallback: 'style-loader',
+                    use: 'css-loader',
+                }),
             },
             {
                 test: /\.(png|jpg|gif|svg|eot|ttf|woff|woff2)$/,
@@ -41,5 +36,5 @@ module.exports = {
             },
         ],
     },
-    plugins: [new MiniCssExtractPlugin({ filename: 'leaflet-geoman.css' }),],
+    plugins: [new ExtractTextPlugin('leaflet.pm.css')],
 };
